docs(README.md): update command line options for dragoman tool
chore(cli.go): update command line options for dragoman tool

The README.md file was updated to reflect the changes in the command line options for the dragoman tool. The `--source-lang` and `--target-lang` options were changed to `--from` and `--to` respectively. The examples in the README.md file were also updated to use the new options.

In the cli.go file, the `SourceLang` and `TargetLang` fields in the `options` struct were updated to use the `name:"from"` and `name:"to"` tags respectively. This change aligns the command line options in the code with the updated options in the README.md file.